src/requirements.txt
python-vlc
configparser
termcolor
rpi.gpio
netifaces